service/Services/entities/src/main/java/org/infy/entities/triggerinputs/TestSteps.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.entities.triggerinputs;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

/**
 * Entity to store test plan list 
 * 
 * @author Infosys
 *
 */
public class TestSteps {

	@SerializedName("stepName")
	@Expose
	private String stepName;

	@SerializedName("toolName")
	@Expose
	private String toolName;

	public String getStepName() {
		return stepName;
	}

	public void setStepName(String stepName) {
		this.stepName = stepName;
	}

	public String getToolName() {
		return toolName;
	}

	public void setToolName(String toolName) {
		this.toolName = toolName;
	}

}
